requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.5
xgboost>=2.1
lightgbm>=4.5
matplotlib>=3.8
seaborn>=0.13
python-dotenv>=1.0
joblib>=1.4
requests>=2.31
